Don't trust PWD; define STRESSMAKE_CWD instead
Stress Make used to take the current directory from the PWD
environment variable passed to it by the server.  Alas, LLVM's
Makefile.config.in rather obnoxiously defines "PWD := @BINPWD@" to set
PWD to, e.g., /bin/pwd.  This was breaking Stress Make, as Stress Make
tried to set the current directory to a non-directory.

In this revision of the code, Stress Make explicitly defines a
STRESSMAKE_CWD environment variable to refer to the current working
directory and passes this from client to server.  So far, this seems
to work around LLVM-style Makefiles that overwrite the PWD variable.